+++ examples/Infrastructure/Model/CourseModel.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Worm\Examples\Infrastructure\Model;
 
+use WoohooLabs\Worm\Examples\Domain\Course;
 use WoohooLabs\Worm\Model\AbstractModel;
 
 class CourseModel extends AbstractModel
@@ -42,13 +43,29 @@ class CourseModel extends AbstractModel
         ];
     }
 
+    public function getClassModel(): ClassModel
+    {
+        return $this->classModel;
+    }
+
+    public function mapCourse(Course $course): array
+    {
+        return [
+            $this->id => $course->getId(),
+            $this->name => $course->getName(),
+            $this->description => $course->getDescription(),
+            $this->credit => $course->getCredit(),
+        ];
+    }
+
     protected function getRelationships(): array
     {
         return [
             "classes" => $this->hasMany(
                 $this->classModel,
                 $this->classModel->course_id,
-                $this->id
+                $this->id,
+                true
             )
         ];
     }